[MBL-19927][Student] Add course home tab screens and refactor packages
Add placeholder screens for each course home tab:
- CourseOverviewScreen (Home tab)
- CourseModulesScreen (Modules tab)
- CourseMyWorkScreen (My Work tab)
- CourseNavigationScreen (More tab)

Each uses LazyColumn to support collapsing toolbar scroll.
Refactor NGC features into features/ package structure.
