+++ libs/ngc/src/main/java/com/instructure/ngc/features/coursehome/CourseHomeViewModel.kt
@@ -14,7 +14,7 @@
  *     limitations under the License.
  */
 
-package com.instructure.ngc.coursehome
+package com.instructure.ngc.features.coursehome
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
