+++ libs/ngc/src/main/java/com/instructure/ngc/features/splash/SplashUiState.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2026 - present Instructure, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ */
+
+package com.instructure.ngc.features.splash
+
+import com.instructure.canvasapi2.models.CanvasTheme
+
+data class SplashUiState(
+    val loading: Boolean = true,
+    val error: Boolean = false,
+    val initialDataLoaded: Boolean = false,
+    val themeToApply: CanvasTheme? = null
+)
